Complete rewrite of role
This is essentially a new role, no backwards compatibility should be assumed with older versions. Note the updated 'overview' in the README as the aims of this role have changed.

Note: This commit represents a 'work in progress' snapshot of rewriting this role. It is not yet complete and is not feature comparable with the existing role (but will be, except where noted in the change log). This snapshot is being committed now to test CI builds.

Most notably, this role now supports CentOS as well as Ubuntu, as per BARC guidelines and adds support for automatically configuring the system firewall on both OSes.

Another notable change is the removal of direct support for generating server blocks (virtual hosts). Instead, this role contains a number of Ansible templates (and role defaults) to generate server blocks as needed, but in tasks external to this role.

Given the range of uses for Nginx, it is not feasible to try to cater to every use-case, or to restrict people to a narrow set of supported use-cases. Instead, if you are using a common use-case you can use a template from this role, otherwise you can generate the required server block yourself without this role interfering.

These changes, and others decisions made in the rewriting this role, stem from the decision to make BARC roles as generic and un-opinionated as is useful. Whilst these changes make this role less 'magic' it also makes it far more flexible for use by others.

In line with all new roles, testing using both CI and manual local methods are used to ensure this role meets its requirements - and typical use-cases in terms of generating common server blocks.