Fixing functional tests in CI
Turns out there is already an Apache server running in the CI environment.

This was sort of a silent error since we use the conventional document root the Apache server was working correctly. In our case we expected visiting over HTTP to result in a 301 redirect to HTTPS, but instead we got served the index file by Apache, not Nginx. If we weren't detecting a non-200 status code this would have been missed completely.

Therefore an extra assertion is added in CI to check the server name is Nginx, and not Apache.