Disabling the global access log
This is largely for performance and cleanliness as we rarely care about this information in apps for example.

This option can now be easily enabled in server-block templates if needed however (though by default it is disabled).

Ideally centralised logging (using tools such as log stash) should be used for logging tasks, though this is not yet provided by BAS Web & Applications (yet).